==> 99_combine_article_types.R <==
# 99_combine_article_types.R
# function to combine articles into a smaller group
# chose top based on frequency and common types
# copied from acronyms
# Jan 2020

combine.types = function(indata){
outdata = mutate(indata,
                type = fct_recode(factor(type), 
                                  'Clinical Trial' = 'Adaptive Clinical Trial',
                                  'Other' = 'Address',
                                  'Other' = 'Autobiography',
                                  'Other' = 'Bibliography',
                                  'Other' = 'Biography',
                                  'Case Reports' = 'Case Reports',
                                  'Other' = 'Classical Article',
                                  'Other' = 'Clinical Conference',
                                  'Other' = 'Clinical Study',
                                  'Clinical Trial' = 'Pragmatic Clinical Trial',
                                  'Clinical Trial' = 'Randomized Controlled Trial, Veterinary',
                                  'Clinical Trial' = 'Clinical Trial',
                                  'Clinical Trial' = 'Clinical Trial, Phase I',
                                  'Clinical Trial' = 'Clinical Trial, Phase II',
                                  'Clinical Trial' = 'Clinical Trial, Phase III',
                                  'Clinical Trial' = 'Clinical Trial, Phase IV',
                                  'Clinical Trial' = 'Clinical Trial Protocol',
                                  'Clinical Trial' = 'Clinical Trial, Veterinary',
                                  'Comment' = 'Comment',
                                  'Other' = 'Collected Works',
                                  'Other' = 'Comparative Study',
                                  'Other' = 'Congress',
                                  'Other' = 'Congresses',
                                  'Other' = 'Consensus Development Conference',
                                  'Other' = 'Consensus Development Conference, NIH',
                                  'Clinical Trial' = 'Controlled Clinical Trial',
                                  'Other' = 'Corrected and Republished Article',
                                  'Other' = 'Dataset',
                                  'Other' = 'Dictionary',
                                  'Other' = 'Directory',
                                  'Other' = 'Duplicate Publication',
                                  'Editorial' = 'Editorial',
                                  'Other' = 'English Abstract',
                                  'Other' = 'Equivalence Trial',
                                  'Other' = 'Evaluation Study',
                                  'Other' = 'Evaluation Studies',
                                  'Other' = 'Expression of Concern',
                                  'Other' = 'Festschrift',
                                  'Other' = 'Government Publication',
                                  'Other' = 'Guideline',
                                  'Other' = 'Historical Article',
                                  'Other' = 'Interactive Tutorial',
                                  'Other' = 'Interview',
                                  'Other' = 'Introductory Journal Article',
                                  'Journal Article' = 'Journal Article',
                                  'Other' = 'Lecture',
                                  'Other' = 'Legal Case',
                                  'Other' = 'Legislation',
                                  'Other' = 'Letter',
                                  'Other' = 'Meta-Analysis',
                                  'Other' = 'Multicenter Study',
                                  'Other' = 'News',
                                  'Other' = 'Newspaper Article',
                                  'Other' = 'Observational Study',
                                  'Other' = 'Overall',
                                  'Other' = 'Patient Education Handout',
                                  'Other' = 'Periodical Index',
                                  'Other' = 'Personal Narrative',
                                  'Other' = 'Practice Guideline',
                                  'Other' = 'Published Erratum',
                                  'Other' = 'Randomized Controlled Trial',
                                  'Other' = 'Research Support, N.I.H., Extramural',
                                  'Other' = 'Research Support, N.I.H., Intramural',
                                  'Other' = "Research Support, Non-U.S. Gov't",
                                  'Other' = "Research Support, U.S. Gov't, Non-P.H.S.",
                                  'Other' = "Research Support, U.S. Gov't, P.H.S.",
                                  'Other' = 'Retracted Publication',
                                  'Other' = 'Retraction of Publication',
                                  'Other' = 'Review',
                                  'Other' = 'Scientific Integrity Review',
                                  'Journal Article' = 'Systematic Review',
                                  'Other' = 'Technical Report',
                                  'Other' = 'Validation Study',
                                  'Other' = 'Video-Audio Media',
                                  'Other' = 'Webcast'),
                type = relevel(type, ref = 'Journal Article')) # journal article is the reference
  return(outdata)
}
